requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1
PyJWT==2.8.0
